Use standard type for potentialAction.target (EntryPoint not URL)
According to Google (nowadays): "Google expands potentialAction.target into urlTemplate. But this is non-standard and Google primarily supports it for Microdata simplicity. For best results, include the whole EntryPoint object."
AFAIK: It seems that Google introduced this feature the wrong way years ago, which is probably why it's still supported by Google today.
=> https://developers.google.com/search/docs/advanced/structured-data/sitelinks-searchbox#website

According to schema.org, potentialAction.target was introduced mid 2014 and the expected type was always "EntryPoint" (never URL or Text).
=> http://web.archive.org/web/20140706031505/http://schema.org/SearchAction